renamed impls subdirs
correct typos
