+++ pyproject.toml
@@ -68,11 +68,11 @@ jedi = "^0.17"
 "smqtk_classifier.impls.classification_element.file" = "smqtk_classifier.impls.classification_element.file"
 "smqtk_classifier.impls.classification_element.memory" = "smqtk_classifier.impls.classification_element.memory"
 "smqtk_classifier.impls.classification_element.postgres" = "smqtk_classifier.impls.classification_element.postgres"
-# Classifier
-"smqtk_classifier.impls.classifier.classify_index_label_descriptor" = "smqtk_classifier.impls.classifier.classify_index_label_descriptor"
+# Classify Descriptor
+"smqtk_classifier.impls.classify_descriptor.classify_index_label_descriptor" = "smqtk_classifier.impls.classify_descriptor.classify_index_label_descriptor"
 # Supervised
-"smqtk_classifier.impls.supervised.libsvm" = "smqtk_classifier.impls.supervised.libsvm"
-"smqtk_classifier.impls.supervised.sklearn_logistic_regression" = "smqtk_classifier.impls.supervised.sklearn_logistic_regression"
+"smqtk_classifier.impls.classify_descriptor_supervised.libsvm" = "smqtk_classifier.impls.classify_descriptor_supervised.libsvm"
+"smqtk_classifier.impls.classify_descriptor_supervised.sklearn_logistic_regression" = "smqtk_classifier.impls.classify_descriptor_supervised.sklearn_logistic_regression"
 
 ###############################################################################
 [tool.pytest.ini_options]
