+++ smqtk_classifier/impls/classify_descriptor/classify_index_label_descriptor.py
