+++ smqtk_classifier/impls/classify_descriptor_supervised/libsvm.py
