+++ smqtk_classifier/impls/classify_descriptor_supervised/sklearn_logistic_regression.py
